Restructure hdf reader and codec filters into a module namespace.
