+++ virtualizarr/readers/hdf/__init__.py
@@ -1,0 +1,11 @@
+from .hdf import (
+    HDFVirtualBackend,
+    construct_virtual_dataset,
+    open_loadable_vars_and_indexes,
+)
+
+__all__ = [
+    "HDFVirtualBackend",
+    "construct_virtual_dataset",
+    "open_loadable_vars_and_indexes",
+]
